refactor: specs and scenarios
